Function sunrise_shell::generate_jobs

pub fn generate_jobs(
    terminal: &mut Terminal, 
    filesystem: &IFileSystemProxy, 
    twili: &ITwiliManagerServiceProxy, 
    loader: &ILoaderInterfaceProxy, 
    line: &str
) -> Result<Vec<Job>, Error>

Generate a vector of Job from a command line.

The returned jobs will not be started, it's up to the caller to start them. The jobs' stdin/stdout/stderr will be properly configured according to the line - pipes and stdin/stdout redirection are supported. Stderr always points to the terminal.