Struct sunrise_shell::Command

struct Command<'a> {
    args: Vec<&'a str>,
    redirect_stdin: Option<&'a str>,
    redirect_stdout: Option<&'a str>,
    pipe_stdout: bool,
}

A command to run as part of a pipeline.

Fields
args: Vec<&'a str>

Array of arguments to pass to the subcommand. First argument is the command name.

redirect_stdin: Option<&'a str>

The filename to take stdin from.

redirect_stdout: Option<&'a str>

The filename to redirect stdout to.

pipe_stdout: bool

True if stdout should be piped to the next command in the cmdline.