Crate sunrise_shell

Shell

Creates an interactive terminal window, providing a few functions useful to test Sunrise. Type help followed by enter to get a list of allowed commands.

Structs

CURRENT_WORK_DIRECTORY
Represent the current work directory.

Statics

CAPABILITIES

CURRENT_WORK_DIRECTORY

HEADER

LOUIS1
Meme for KFS1

LOUIS2
Meme for KFS2

LOUIS3
Meme for KFS3

LOUIS4
Meme for KFS4

LOUIS5
Meme for KFS5

LOUIS6
Meme for KFS6

LOUIS7
Meme for KFS7

Functions

cat
Print a file on the standard output.

cd
Change the current working directory

get_absolute_path
Get an absolute path from an user path

get_next_line
Read key presses until a \n is detected, and return the string (excluding \n). The key presses should be written to stdout.

get_next_line_no_echo
Read key presses until a \n is detected, and return the string (excluding \n). Don't print the key presses on stdout.

get_path_relative_to_current_directory
Get a path relative to the current directory

login
Asks the user to login repeatedly. Returns with an error if the /etc/passwd file is invalid or doesn't exist.

ls
List files and folders at the given path, or in the current path is none is given.

main

show_gif
Shows a GIF in a new window, blocking the caller. When a key is pressed, the window is closed and control is given back to the caller.

split_path
Splits a path at the first / it encounters.

test_divide_by_zero
Test function ensuring divide by zero interruption kills only the current process.

test_page_fault
Test function ensuring pagefaults kills only the current process.

test_threads
Test function ensuring threads are working properly.

user_add
Adds a new user to /etc/passwd with the specified username.